requirements.txt
fastapi
uvicorn
python-jose
python-dotenv
email-validator
pymysql
sqlalchemy
mysql-connector-python
requests
query 